pyproject.toml
[tool.poetry]
name = "modernapi"
version = "0.1.0"
description = "Building a secured API with FastAPI"
authors = ["Prayson W. Daniel <praysonpi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.68.2"
sqlmodel = "^0.0.4"
uvicorn = "^0.15.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.5"
bcrypt = "^3.2.0"
python-decouple = "^3.5"

[tool.poetry.dev-dependencies]
black = "flake8"
flake8 = "^3.9.2"
pytest = "^6.2.5"
pylint = "^2.11.1"
invoke = "^1.6.0"
requests = "^2.27.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
